fix: Family, Comment Redis를 이용한 유효 사용자 인증 및 인가

+++ family-moments/src/main/java/com/spring/familymoments/domain/comment/CommentController.java
@@ -4,6 +4,7 @@ import com.spring.familymoments.config.BaseException;
 import com.spring.familymoments.config.BaseResponse;
 import com.spring.familymoments.domain.comment.model.GetCommentsRes;
 import com.spring.familymoments.domain.comment.model.PostCommentReq;
+import com.spring.familymoments.domain.user.AuthService;
 import com.spring.familymoments.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -11,12 +12,15 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+import static com.spring.familymoments.config.BaseResponseStatus.INVALID_JWT;
+
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/comments")
 public class CommentController {
     private final CommentService commentService;
+    private final AuthService authService;
 
     /**
      * 댓글 생성 API
@@ -26,8 +30,14 @@ public class CommentController {
     @ResponseBody
     @PostMapping("")
     public BaseResponse<String> createComment(@AuthenticationPrincipal User user,
-                                             @RequestParam("postId") Long postId,
-                                              @RequestPart PostCommentReq postCommentReq){
+                                              @RequestParam("postId") Long postId,
+                                              @RequestPart PostCommentReq postCommentReq,
+                                              @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try{
             commentService.createComment(user, postId, postCommentReq);
             return new BaseResponse<>("댓글이 업로드되었습니다.");
@@ -43,7 +53,12 @@ public class CommentController {
      */
     @ResponseBody
     @GetMapping("")
-    public BaseResponse<List<GetCommentsRes>> getCommentsByPostId(@RequestParam("postId") Long postId){
+    public BaseResponse<List<GetCommentsRes>> getCommentsByPostId(@RequestParam("postId") Long postId,
+                                                                  @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try{
             List<GetCommentsRes> getCommentsRes = commentService.getCommentsByPostId(postId);
             return new BaseResponse<>(getCommentsRes);
@@ -59,7 +74,13 @@ public class CommentController {
      */
     @ResponseBody
     @DeleteMapping("/{commentId}")
-    public BaseResponse<String> deleteComment(@AuthenticationPrincipal User user, @PathVariable Long commentId){
+    public BaseResponse<String> deleteComment(@AuthenticationPrincipal User user,
+                                              @PathVariable Long commentId,
+                                              @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try{
             commentService.deleteComment(user, commentId);
             return new BaseResponse<>("댓글이 삭제되었습니다.");
