+++ family-moments/src/main/java/com/spring/familymoments/domain/family/FamilyController.java
@@ -26,9 +26,9 @@ public class FamilyController {
 
     private final FamilyService familyService;
     private final JwtService jwtService;
+    private final AuthService authService;
     @Autowired
     private final AwsS3Service awsS3Service;
-    private final AuthService authService;
 
 
     /**
@@ -40,7 +40,13 @@ public class FamilyController {
     @PostMapping("/family")
     public BaseResponse<PostFamilyRes> createFamily(@AuthenticationPrincipal User user,
                                                     @RequestParam(name = "representImg") MultipartFile representImg,
-                                                    @RequestPart PostFamilyReq postFamilyReq) {
+                                                    @RequestPart PostFamilyReq postFamilyReq,
+                                                    @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) {
+
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try{
 //        int owner = jwtService.getUserIdx();
             // 대표 이미지 넣기
@@ -88,7 +94,12 @@ public class FamilyController {
     @ResponseBody
     @GetMapping("/{familyId}/created")
     public BaseResponse<GetFamilyCreatedNicknameRes> getFamilyCreatedNickname(@AuthenticationPrincipal User user,
-                                                                              @PathVariable Long familyId){
+                                                                              @PathVariable Long familyId,
+                                                                              @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try {
             GetFamilyCreatedNicknameRes getFamilyCreatedNicknameRes = familyService.getFamilyCreatedNickname(user, familyId);
             return new BaseResponse<>(getFamilyCreatedNicknameRes);
@@ -103,7 +114,12 @@ public class FamilyController {
      * @return BaseResponse<FamilyDto>
      */
     @GetMapping("/{familyId}/users")
-    public BaseResponse<List<GetFamilyAllRes>> getFamilyAll(@PathVariable Long familyId){
+    public BaseResponse<List<GetFamilyAllRes>> getFamilyAll(@PathVariable Long familyId,
+                                                            @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try {
             List<GetFamilyAllRes> getFamilyAllRes = familyService.getFamilyAll(familyId);
             return new BaseResponse<>(getFamilyAllRes);
@@ -196,7 +212,12 @@ public class FamilyController {
     @PatchMapping("/{familyId}")
     public BaseResponse<String> updateUploadCycle(@AuthenticationPrincipal User user,
                                                   @PathVariable Long familyId,
-                                                  @RequestParam("uploadCycle") int uploadCycle){
+                                                  @RequestParam("uploadCycle") int uploadCycle,
+                                                  @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try {
             familyService.updateUploadCycle(user, familyId, uploadCycle);
             return new BaseResponse<>("업로드 주기가 수정되었습니다.");
@@ -211,7 +232,13 @@ public class FamilyController {
      * @return BaseResponse<String>
      */
     @DeleteMapping("/{familyId}")
-    public BaseResponse<String> deleteFamily(@AuthenticationPrincipal User user, @PathVariable Long familyId) {
+    public BaseResponse<String> deleteFamily(@AuthenticationPrincipal User user,
+                                             @PathVariable Long familyId,
+                                             @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) {
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try {
             familyService.deleteFamily(user, familyId);
             return new BaseResponse<>("가족이 삭제되었습니다.");
